embedding/requirements.txt
fastapi
uvicorn[standard]
requests
torch
Pillow
open-clip-torch
sentence-transformers
numpy
